+++ spec/functional/mongoid/named_scope_spec.rb
@@ -90,5 +90,37 @@ describe Mongoid::NamedScope do
       end
 
     end
+    
+    context "when overwriting an existing scope" do
+
+      it "logs warnings per default" do
+        require 'stringio'
+        log_io = StringIO.new
+        Mongoid.logger = ::Logger.new(log_io)
+        Mongoid.scope_overwrite_exception = false
+        
+        Person.class_eval do
+          scope :old, criteria.where(:age.gt => 67)
+        end
+        
+        log_io.rewind
+        log_io.readlines.join.should =~
+          /Creating scope :old. Overwriting existing method Person.old/
+      end
+      
+      it "throws exception if configured with scope_overwrite_exception = true" do
+        Mongoid.scope_overwrite_exception = true
+        lambda {
+          Person.class_eval do
+            scope :old, criteria.where(:age.gt => 67)
+          end
+        }.should raise_error(
+          Mongoid::Errors::ScopeOverwrite,
+          "Cannot create scope :old, because of existing method Person.old."
+        )
+      end
+
+    end
+
   end
 end